+++ love/src/components/AuxTel/Camera/Camera.jsx
@@ -66,6 +66,11 @@ export default class Camera extends Component {
   };
 
   render() {
+    const imageIds = Object.keys(this.props.imageSequence?.images ?? {});
+    const orederedImages = imageIds.sort((a, b) => {
+      return this.props.imageSequence.images[b].timestamp - this.props.imageSequence.images[a].timestamp;
+    });
+
     return (
       <div className={styles.cameraContainer}>
         <div className={styles.statesContainer}>
@@ -107,8 +112,8 @@ export default class Camera extends Component {
                 </tr>
               </thead>
               <tbody>
-                {this.props.imageSequence.images &&
-                  Object.keys(this.props.imageSequence.images).map((imageName) => {
+                {orederedImages &&
+                  orederedImages.map((imageName) => {
                     const image = this.props.imageSequence.images[imageName];
                     const imageKey = `${this.props.imageSequence.name}-${imageName}`;
                     const isIntegrating = image.state === 'INTEGRATING';