+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Version History
 ===============
 
+v5.18.0
+--------
+
+* Add Location Control sources of information `<https://github.com/lsst-ts/LOVE-frontend/pull/446>`_
+* Fix LATISS and ATCamera state mappings `<https://github.com/lsst-ts/LOVE-frontend/pull/445>`_
+* Add ``Aircraft Tracker`` component implementation `https://github.com/lsst-ts/LOVE-frontend/pull/447`
+
 v5.17.1
 --------